Bacteria
Updated Bacteria for more variation in movement . 
Created game manager script and made it so a set number of bacteria will be spawned over a set period of time. This avoids cluttering of enemies at the beginning. Bacteria are spawned both inside and outside the mouth simultaneously.

Note: create panel at start of game that tells player to "Press R to switch to Inside/Outside of mouth